+++ src/org/camunda/latera/bss/helpers/hydra/Service.groovy
@@ -200,6 +200,7 @@ trait Service {
     def equipmentId        = order."${equipmentPrefix}Id"
     def parSubscriptionId  = order."${parSubscriptionPrefix}Id"
     def prevSubscriptionId = order."${prevSubscriptionPrefix}Id"
+
     Map subscription = hydra.createSubscription(
       customerId,
       accountId          : accountId,
@@ -208,7 +209,7 @@ trait Service {
       equipmentId        : equipmentId,
       beginDate          : params.beginDate,
       endDate            : params.endDate,
-      invoiceEndDate     : params.endDate,
+      chargeLogEndDate   : params.endDate,
       payDay             : params.payDay,
       parSubscriptionId  : parSubscriptionId,
       prevSubscriptionId : prevSubscriptionId